💥 Move backend protocols from abc to _protocol modules
