+++ docs/testing.md
@@ -13,7 +13,7 @@ from unittest.mock import AsyncMock
 
 from grelmicro import Grelmicro
 from grelmicro.coordination import Coordination
-from grelmicro.coordination.abc import LockBackend
+from grelmicro.coordination import LockBackend
 
 
 async def test_login(micro: Grelmicro) -> None:
