+++ grelmicro/clock/__init__.py
@@ -5,10 +5,10 @@ real clock with no setup. Tests install a `VirtualClock` to drive time-dependent
 behavior without real waiting.
 """
 
+from grelmicro.clock._protocol import ClockBackend
 from grelmicro.clock._real import RealClock
 from grelmicro.clock._seam import monotonic, sleep
-from grelmicro.clock.abc import ClockBackend
-from grelmicro.clock.virtual import VirtualClock
+from grelmicro.clock._virtual import VirtualClock
 
 __all__ = [
     "ClockBackend",
