+++ grelmicro/config/__init__.py
@@ -1,7 +1,7 @@
 """Externalized configuration: reconfigure live components from a source."""
 
 from grelmicro.config._external import ExternalConfig
-from grelmicro.config.abc import ConfigBackend
+from grelmicro.config._protocol import ConfigBackend
 from grelmicro.config.file import FileConfigAdapter
 
 __all__ = [
