+++ grelmicro/coordination/postgres.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Annotated, Any, ClassVar, Self
 
 from typing_extensions import Doc
 
-from grelmicro.coordination.abc import (
+from grelmicro.coordination._protocol import (
     LeaderRecord,
     LockBackend,
     ScheduleBackend,
